# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kolibri AI OS V2.1",
    "version": "2.1.0"
  },
  "paths": {
...
    "/v1/responses": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Create Response",
        "operationId": "create_response_v1_responses_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/{response_id}": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "Get Response",
        "operationId": "get_response_v1_responses__response_id__get",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          },
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Stream"
            }
          },
          {
            "name": "starting_after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Starting After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "delete": {
        "tags": [
          "responses"
        ],
        "summary": "Delete Response",
        "operationId": "delete_response_v1_responses__response_id__delete",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/{response_id}/events": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "Response Events",
        "operationId": "response_events_v1_responses__response_id__events_get",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          },
          {
            "name": "starting_after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Starting After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/{response_id}/cancel": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Cancel Response",
        "operationId": "cancel_response_v1_responses__response_id__cancel_post",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/{response_id}/input_items": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "Input Items",
        "operationId": "input_items_v1_responses__response_id__input_items_get",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/input_tokens": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Input Tokens",
        "operationId": "input_tokens_v1_responses_input_tokens_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/responses/compact": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Compact Response",
        "operationId": "compact_response_v1_responses_compact_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/models": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "List Models",
        "operationId": "list_models_v1_models_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/models/{model_id}": {
      "get": {
        "tags": [
          "responses"
        ],
        "summary": "Get Model",
        "operationId": "get_model_v1_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      },
      "delete": {
        "tags": [
          "responses"
        ],
        "summary": "Delete Model",
        "operationId": "delete_model_v1_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "responses"
        ],
        "summary": "Chat Completions",
        "operationId": "chat_completions_v1_chat_completions_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "KolibriSessionCookie": []
          },
          {
            "KolibriSessionHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCompletionCreate": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "kolibri"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "title": "Messages"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "messages"
        ],
        "title": "ChatCompletionCreate"
      },
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseCreate": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "kolibri"
          },
          "input": {
            "title": "Input"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "conversation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation"
          },
          "previous_response_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Response Id"
          },
          "background": {
            "type": "boolean",
            "title": "Background",
            "default": false
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "default": false
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Tools"
          },
          "tool_choice": {
            "title": "Tool Choice",
            "default": "auto"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "input"
        ],
        "title": "ResponseCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk-kolibri API key or signed Kolibri session",
        "description": "OpenAI-compatible bearer authentication."
      },
      "KolibriSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kolibri_session",
        "description": "Secure HttpOnly browser session cookie."
      },
      "KolibriSessionHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Session"
      },
      "KolibriNodeToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Kolibri-Node-Token"
      }
    }
  },
...
}